+++ lib/decorators/index.ts
@@ -1,5 +1,3 @@
-export * from './loader-context.decorator';
-export * from './loader-queries.decorator';
 export * from './resolve-loader.decorator';
 export * from './validation-rule.decorator';
 export * from './resolve-reference-loader.decorator';