+++ lib/mercurius-core.module.ts
@@ -1,0 +1,9 @@
+import { Global, Module } from '@nestjs/common';
+import { PubSubHost } from './services';
+
+@Global()
+@Module({
+  providers: [PubSubHost],
+  exports: [PubSubHost],
+})
+export class MercuriusCoreModule {}