+++ tests/e2e/code-first.spec.ts
@@ -1,23 +1,49 @@
-import { suite } from 'uvu';
-import * as assert from 'uvu/assert';
-import { Test } from '@nestjs/testing';
 import {
   FastifyAdapter,
   NestFastifyApplication,
 } from '@nestjs/platform-fastify';
-import { ApplicationModule } from '../code-first/app.module';
 import { createTestClient } from '../utils/create-test-client';
+import { suite } from 'uvu';
+import * as assert from 'uvu/assert';
+import { Test } from '@nestjs/testing';
+import { AppModule } from '../code-first/app.module';
 
 interface Context {
   app: NestFastifyApplication;
   mercuriusClient: ReturnType<typeof createTestClient>;
 }
 
-const graphqlSuite = suite<Context>('Code-first');
+const expectedCats = [
+  {
+    id: '1',
+    species: 'CAT',
+    lives: 9,
+  },
+  {
+    id: '2',
+    species: 'CAT',
+    lives: 9,
+  },
+];
 
-graphqlSuite.before.each(async (ctx) => {
+const expectedDogs = [
+  {
+    id: '1',
+    species: 'DOG',
+    age: 2,
+  },
+  {
+    id: '2',
+    species: 'DOG',
+    age: 5,
+  },
+];
+
+const gqlSuite = suite<Context>('Code-first');
+
+gqlSuite.before.each(async (ctx) => {
   const module = await Test.createTestingModule({
-    imports: [ApplicationModule],
+    imports: [AppModule],
   }).compile();
 
   const app = module.createNestApplication<NestFastifyApplication>(
@@ -28,145 +54,152 @@ graphqlSuite.before.each(async (ctx) => {
   ctx.mercuriusClient = createTestClient(module);
 });
 
-graphqlSuite.after.each(async (ctx) => {
+gqlSuite.after.each(async (ctx) => {
   await ctx.app.close();
 });
 
-graphqlSuite(
-  'should return the categories result',
+gqlSuite('should return a list of cats', async ({ mercuriusClient }) => {
+  const response = await mercuriusClient.query(`
+    query cats{
+      cats {
+        id
+        species
+        lives
+      }
+    }
+  `);
+
+  assert.equal(response.data, {
+    cats: expectedCats,
+  });
+});
+
+gqlSuite('should return Loader values', async ({ mercuriusClient }) => {
+  const response = await mercuriusClient.query(
+    `
+    query cat($id: ID!) {
+      cat(id: $id) {
+        hasFur
+      }    
+    }
+  `,
+    {
+      variables: { id: 1 },
+    },
+  );
+
+  assert.equal(response.data, {
+    cat: {
+      hasFur: true,
+    },
+  });
+});
+
+gqlSuite('should apply schema directives', async ({ mercuriusClient }) => {
+  const response = await mercuriusClient.query(
+    `
+    query cat($id: ID!) {
+      cat(id: $id) {
+        name
+      }    
+    }
+  `,
+    {
+      variables: { id: 1 },
+    },
+  );
+
+  assert.equal(response.data, {
+    cat: {
+      name: 'FUFY',
+    },
+  });
+});
+
+gqlSuite(
+  'should return interface fieldResolver value',
   async ({ mercuriusClient }) => {
     const response = await mercuriusClient.query(
-      `query {
-      categories {
-        name
-        description
-        tags
-      }
-    }`,
+      `
+    query cat($id: ID!) {
+      cat(id: $id) {
+        hasPaws
+      }    
+    }
+  `,
+      {
+        variables: { id: 1 },
+      },
     );
 
     assert.equal(response.data, {
-      categories: [
-        {
-          name: 'Category #1',
-          description: 'default value',
-          tags: [],
-        },
-      ],
+      cat: {
+        hasPaws: true,
+      },
     });
   },
 );
 
-graphqlSuite('should return the search result', async ({ mercuriusClient }) => {
-  const response = await mercuriusClient.query(
-    `query {
-      search {
-        ... on Recipe {
-          title
+gqlSuite('should return union', async ({ mercuriusClient }) => {
+  const response = await mercuriusClient.query(`
+    query {
+      domesticAnimals {
+        ... on Animal {
+          id
+          species
+          hasPaws
         }
-        ... on Ingredient {
-          name
+        ...on Dog {
+          age
+        }
+        ... on Cat {
+          lives
         }
       }
-    }`,
-  );
+    }
+  `);
 
   assert.equal(response.data, {
-    search: [
-      {
-        title: 'recipe',
-      },
-      {
-        name: 'test',
-      },
+    domesticAnimals: [
+      ...expectedDogs.map((d) => ({
+        ...d,
+        hasPaws: true,
+      })),
+      ...expectedCats.map((c) => ({
+        ...c,
+        hasPaws: true,
+      })),
     ],
   });
 });
 
-graphqlSuite(`should return query result`, async ({ mercuriusClient }) => {
-  const response = await mercuriusClient.query(
-    `query {
-      recipes {
-        id
-        description
-        ingredients {
-          name
+gqlSuite('should filter union return', async ({ mercuriusClient }) => {
+  const response = await mercuriusClient.query(`
+    query {
+      domesticAnimals(species: DOG) {
+        ... on Animal {
+          id
+          species
+          hasPaws
         }
-        rating
-        interfaceResolver
-        averageRating
+        ...on Dog {
+          age
+        }
+        ... on Cat {
+          lives
+        }
       }
-    }`,
-  );
+    }
+  `);
 
   assert.equal(response.data, {
-    recipes: [
-      {
-        id: '1',
-        description: 'Description: Calzone',
-        ingredients: [
-          {
-            name: 'cherry',
-          },
-        ],
-        rating: 10,
-        interfaceResolver: true,
-        averageRating: 0.5,
-      },
-      {
-        id: '2',
-        description: 'Placeholder',
-        ingredients: [
-          {
-            name: 'cherry',
-          },
-        ],
-        rating: 10,
-        interfaceResolver: true,
-        averageRating: 0.5,
-      },
+    domesticAnimals: [
+      ...expectedDogs.map((d) => ({
+        ...d,
+        hasPaws: true,
+      })),
     ],
   });
 });
 
-graphqlSuite(`should return query result`, async ({ mercuriusClient }) => {
-  const response = await mercuriusClient.query(
-    `query {
-      recipes {
-        id
-        ingredients {
-          name
-        }
-        rating
-        averageRating
-      }
-    }`,
-  );
-
-  assert.equal(response.data, {
-    recipes: [
-      {
-        id: '1',
-        ingredients: [
-          {
-            name: 'cherry',
-          },
-        ],
-        rating: 10,
-        averageRating: 0.5,
-      },
-      {
-        id: '2',
-        ingredients: [
-          {
-            name: 'cherry',
-          },
-        ],
-        rating: 10,
-        averageRating: 0.5,
-      },
-    ],
-  });
-});
-
-graphqlSuite.run();
+gqlSuite.run();