+++ tests/example/modules/complexity/config.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class ConfigService {
+  get maxComplexity() {
+    return 100;
+  }
+}