+++ tests/example/modules/user/inputs/create-user.input.ts
@@ -1,0 +1,5 @@
+import { InputType, OmitType } from '@nestjs/graphql';
+import { UserType } from '../../../types/user.type';
+
+@InputType()
+export class CreateUserInput extends OmitType(UserType, ['id'], InputType) {}