+++ tests/example/modules/user/inputs/full-name.args.ts
@@ -1,0 +1,7 @@
+import { ArgsType, Field } from '@nestjs/graphql';
+
+@ArgsType()
+export class FullNameArgs {
+  @Field({ nullable: true })
+  filter?: string;
+}