+++ tests/example/types/user.type.ts
@@ -1,16 +1,20 @@
-import { Field, ID, ObjectType } from '@nestjs/graphql';
+import { Directive, Extensions, Field, ID, ObjectType } from '@nestjs/graphql';
 import { HashScalar } from '../scalars/hash.scalar';
 import { JSONResolver } from 'graphql-scalars';
+import { Person } from './person.interface';
 
-@ObjectType('User')
-export class UserType {
+@ObjectType('User', {
+  implements: [Person],
+})
+export class UserType implements Person {
   @Field(() => ID)
   id: string;
 
+  @Extensions({ role: 'ADMIN' })
   @Field(() => String, { nullable: true })
   name?: string;
 
-  @Field({ nullable: true })
+  @Field({ defaultValue: 'noone' })
   lastName?: string;
 
   @Field(() => Date)