+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# [0.4.0](https://github.com/Davide-Gheri/nestjs-mercurius/compare/0.3.0...0.4.0) (2021-02-19)
+
+
+### Features
+
+* **pubsub:** global pubsub host' ([064640b](https://github.com/Davide-Gheri/nestjs-mercurius/commit/064640b5c64a68c21456810864bfcf59d7c2e76c))
+
+# [0.3.0](https://github.com/Davide-Gheri/nestjs-mercurius/compare/0.2.0...0.3.0) (2021-02-18)
+
+
+### Features
+
+* add opts object to loaders ([2a94abf](https://github.com/Davide-Gheri/nestjs-mercurius/commit/2a94abfd6bca0f0f6b16a96da0d19e3b09836c4f))
+* ResolveLoader accepts optional opts object ([5ad6dfe](https://github.com/Davide-Gheri/nestjs-mercurius/commit/5ad6dfeb45b9213c1fbc041fe0be236d679ba32f))
+
+# [0.2.0](https://github.com/Davide-Gheri/nestjs-mercurius/compare/0.1.0...0.2.0) (2021-02-18)
+
+
+### Features
+
+* loader middleware, testing unions/enums/interfaces ([35e25e3](https://github.com/Davide-Gheri/nestjs-mercurius/commit/35e25e384a8951a80816ada6f72ad0c3d323a56b))
+
+# [0.1.0](https://github.com/Davide-Gheri/nestjs-mercurius/compare/0.1.0-beta.1...0.1.0) (2021-02-17)
+
 * test: resolvers merging with third party resolvers (58184dc)
 * Merge pull request #4 from Davide-Gheri/feature/query-complexity (881952f)
 * ci: run only on 1 os (5b740c4)